feat: resolving manifests / apps status at runtime instead of at compilation time
