+++ src/DealroadshowK8SBundle.php
@@ -7,7 +7,6 @@ namespace Dealroadshow\Bundle\K8SBundle;
 use Dealroadshow\Bundle\K8SBundle\DependencyInjection\Compiler\AppsPass;
 use Dealroadshow\Bundle\K8SBundle\DependencyInjection\Compiler\AutoSetReplicasPass;
 use Dealroadshow\Bundle\K8SBundle\DependencyInjection\Compiler\AutoSetResourcesPass;
-use Dealroadshow\Bundle\K8SBundle\DependencyInjection\Compiler\EnabledManifestsPass;
 use Dealroadshow\Bundle\K8SBundle\DependencyInjection\Compiler\LocalizationStrategyPass;
 use Dealroadshow\Bundle\K8SBundle\DependencyInjection\Compiler\ManifestGeneratorContextsPass;
 use Dealroadshow\Bundle\K8SBundle\DependencyInjection\Compiler\ManifestsPass;
@@ -37,7 +36,6 @@ class DealroadshowK8SBundle extends Bundle
 
         $container
             ->addCompilerPass(new AppsPass())
-            ->addCompilerPass(pass: new EnabledManifestsPass(), priority: -8)
             ->addCompilerPass(pass: new SetAppConfigPass(), type: PassConfig::TYPE_OPTIMIZE, priority: -8)
             ->addCompilerPass(pass: new ManifestsPass(), type: PassConfig::TYPE_OPTIMIZE, priority: -16)
             ->addCompilerPass(pass: new ManifestGeneratorContextsPass())
